Buildsystem enhanced Cross compiler toolchain building
git-svn-id: https://svn.galauner.de/GeexOS@14 5530181a-2ae2-423d-a44f-b903e3aa7be7

+++ trunk/kernel/inc/size_t.h
@@ -20,7 +20,7 @@ extern "C"
 #endif
 
 /* standard size_t type */
-typedef unsigned size_t;
+typedef long unsigned int size_t;
 
 #ifdef __cplusplus
 }
